requirements.txt
flask>=2.0.0
flask-cors>=6.0.0
python-dotenv>=0.19.0
requests>=2.26.0
pytest>=7.0.0
stripe>=14.1.0
PyJWT>=2.9.0
pydantic>=2.6.3
flask-limiter>=3.5.0
structlog>=25.5.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
gunicorn>=21.2.0
waitress>=3.0.0
